requirements.txt
sounddevice
numpy
openai-whisper